requirements.txt
azure-core==1.24.0
azure-identity==1.10.0
azure-storage-blob==12.12.0
azure-mgmt-storage==20.0.0
azure-mgmt-costmanagement==3.0.0
tabulate==0.8.9
pandas==1.4.2